requirement.txt
anyio==3.6.2
arrow==1.2.3
asyncio-dgram==2.1.2
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookiecutter==1.7.3
dnspython==2.2.1
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
MarkupSafe==2.1.1
mcstatus==10.0.1
msgpack==1.0.4
multidict==6.0.3
nb-cli==0.6.8
nonebot-adapter-onebot==2.1.5
nonebot-plugin-guild-patch==0.2.1
nonebot2==2.0.0rc2
pip==22.3.1
poyo==0.5.0
prompt-toolkit==3.0.36
pydantic==1.10.2
pyfiglet==0.8.post1
pygtrie==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
text-unidecode==1.3
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
wheel==0.37.1
win32-setctime==1.1.0
wincertstore==0.2
yarl==1.8.2
